feat(api): Add async protocol method for request interceptors IAM signing (#2871)
